+++ hedgehog-plutus-simple.cabal
@@ -92,9 +92,11 @@ library
   import:          common-lang
   exposed-modules:
     Hedgehog.Plutus.Adjunction
+    Hedgehog.Plutus.Model
+    Hedgehog.Plutus.Model.Internal
+    Hedgehog.Plutus.Model.TxHelpers
     Hedgehog.Plutus.Simple
     Hedgehog.Plutus.TestSingleScript
-    Hedgehog.Plutus.Tx
     Hedgehog.Plutus.TxTest
 
   build-depends:
