+++ src/Hedgehog/Plutus/Model.hs
@@ -1,0 +1,57 @@
+{-# OPTIONS_GHC -Wno-missing-import-lists #-}
+
+-- I think they are fine here because they are just reexported
+
+module Hedgehog.Plutus.Model (
+  Tx (..),
+  BalancedTx,
+  TxIn (..),
+  TxOut (..),
+  Script (..),
+  InScript (..),
+  ScriptSource (..),
+  TxContext (..),
+  ScriptPurpose (..),
+  Spend (..),
+  balanceTx,
+  forgetBlanced,
+  balanceTxAsPubKey,
+  balanceTxWhere,
+  confirmBalanced,
+  spendWhere,
+  payToUser,
+  payToScript,
+  spendPubKeyUtxo,
+  spendScriptOutput,
+  mintWithRed,
+  scriptPurposeTx,
+)
+where
+
+import Hedgehog.Plutus.Model.Internal (
+  BalancedTx,
+  InScript (..),
+  Script (..),
+  ScriptPurpose (..),
+  ScriptSource (..),
+  Spend (..),
+  Tx (..),
+  TxContext (..),
+  TxIn (..),
+  TxOut (..),
+  balanceTx,
+  balanceTxAsPubKey,
+  balanceTxWhere,
+  confirmBalanced,
+  forgetBlanced,
+  scriptPurposeTx,
+  spendWhere,
+ )
+
+import Hedgehog.Plutus.Model.TxHelpers (
+  mintWithRed,
+  payToScript,
+  payToUser,
+  spendPubKeyUtxo,
+  spendScriptOutput,
+ )
